Reorganise code base
