+++ src/graphics/mod.rs
@@ -1,1 +1,2 @@
 pub mod dimensions;
+pub mod shaders;
